+++ prajna/transform/utility.hpp
@@ -115,7 +115,7 @@ inline std::list<std::shared_ptr<ir::Value>> CaptureExternalValueInClosure(
 
     utility::EachValue(ir_function, [&](std::shared_ptr<ir::Value> ir_value) {
         if (auto ir_instruction = Cast<ir::Instruction>(ir_value)) {
-            for (size_t i = 0; i < ir_instruction->OperandSize(); ++i) {
+            for (int64_t i = 0; i < ir_instruction->OperandSize(); ++i) {
                 auto ir_operand = ir_instruction->GetOperand(i);
                 if (ir_operand->is_global) continue;
                 if (ir_operand->GetParentFunction() != ir_function) {
@@ -136,7 +136,7 @@ inline std::list<std::shared_ptr<ir::Variable>> CaptureExternalVariablesInBlock(
         // 先判断是否是ir::For, 因为ir::For本身也是ir::Instruction
         PRAJNA_ASSERT(!Is<ir::For>(ir_value));
         if (auto ir_instruction = Cast<ir::Instruction>(ir_value)) {
-            for (size_t i = 0; i < ir_instruction->OperandSize(); ++i) {
+            for (int64_t i = 0; i < ir_instruction->OperandSize(); ++i) {
                 auto ir_operand = ir_instruction->GetOperand(i);
                 if (ir_operand->parent_block != ir_block) {
                     if (auto ir_local_variable = Cast<ir::Variable>(ir_operand)) {
